Prevent moves being always merged in the same groups
If we have a unique transfer for 2 sales orders, we have one joint
procurement group for them. If we delivery the first sale order, we
expect to no longer have the first SO in the group of the backorder.

This will prevent to have a never-ending group with all the orders from
the beginning if we have many backorders.

To so do, we consider that a "printed" or done transfer will no longer
change and therefore, we consider its group frozen.